fix(grokbot): split hub listing helpers to stay under the module ceiling
Move hub-authority listing, projection, and report-pairing helpers into
grokbot_jobs_hub and re-export the same names from grokbot_jobs so
callers and tests keep their existing bindings.
